Refactor ProductImage component to use MediaFragment and improve image loading handling
